Apply Gene Set Enrichment Analysis to Multiple Phenotypes

This function applies fast preranked gene set enrichment analysis (FGSEA) to multiple phenotypes using a provided gene set. It is designed to handle large-scale datasets efficiently.

Arguments

input_geneset

A list containing one gene set for enrichment analysis. The element of the list should be a character vector of gene Ensembl IDs.

multiple_phenotypes

A data frame containing multiple phenotypes from Genebass.

...

Additional arguments passed to ukbfgsea.

Value

A data frame containing the results of ukbfgsea for each phenotype annotation. This includes enrichment analysis outcomes using two different ranking metrics, and, if specified, enrichment plots.

Examples

if (FALSE) { # \dontrun{
  library(tidyverse)
  library(ukbFGSEA)
  asd_output <- map_ukbfgsea(input_geneset = ASD185, multiple_phenotypes = genebass_test_data)
} # }